Refator unit testing of plugins
